+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+# 3.2.4 (April 23, 2025)
+
+## Notes
+- Golang: **v1.22**
+
+### Bug Fixes
+
+[PR #323](https://github.com/zscaler/zscaler-sdk-go/pull/323) – Fixed ZPA resource `pracredentialpool` missing attribute `credentialMappingCount`
+[PR #323](https://github.com/zscaler/zscaler-sdk-go/pull/323) – Fixed ZPA resource `policysetcontrollerv2` attribute `credentialPool` pointer
+
 # 3.2.3 (April 22, 2025)
 
 ## Notes